requirements.txt
google-genai
pyyaml
requests
markdown
reportlab
feedparser
newspaper3k
lxml_html_clean
python-dotenv